requirements.txt
yfinance==0.2.61
tti==0.2.2
backtesting==0.6.4
backtrader==1.9.78
pandas_datareader
matplotlib
vectorbt
pandas-ta
ta